Extract ResetPersistStatement to reduce duplicate logic for MySQL and doris
