+++ parser/sql/statement/src/main/java/org/apache/shardingsphere/sql/parser/sql/dialect/statement/mysql/dal/MySQLResetPersistStatement.java
@@ -17,21 +17,16 @@
 
 package org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.sql.parser.sql.common.statement.AbstractSQLStatement;
-import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.DALStatement;
+import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.ResetPersistStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.value.identifier.IdentifierValue;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.MySQLStatement;
 
 /**
  * MySQL reset persist statement.
  */
-@RequiredArgsConstructor
-@Getter
-public final class MySQLResetPersistStatement extends AbstractSQLStatement implements DALStatement, MySQLStatement {
+public final class MySQLResetPersistStatement extends ResetPersistStatement implements MySQLStatement {
     
-    private final boolean ifExists;
-    
-    private final IdentifierValue identifier;
+    public MySQLResetPersistStatement(final boolean ifExists, final IdentifierValue identifier) {
+        super(ifExists, identifier);
+    }
 }
